+++ src/constants/cssAtRulesWithDeclarations/index.js
@@ -1,0 +1,7 @@
+export const cssAtRulesWithDeclarations = [
+	'document',
+	'keyframes',
+	'media',
+	'page',
+	'supports',
+];