+++ src/constants/cssDeprecatedMediaFeatures/index.js
@@ -1,0 +1,9 @@
+/**
+ * https://drafts.csswg.org/mediaqueries/#descdef-media-update
+ */
+
+export const cssDeprecatedMediaFeatures = [
+	'device-aspect-ratio',
+	'device-height',
+	'device-width',
+];