+++ src/constants/cssDeprecatedMediaTypes/index.js
@@ -1,0 +1,24 @@
+/**
+ * https://developer.mozilla.org/en-US/docs/Web/CSS/Media_Queries/Using_media_queries#Media_types
+ * https://drafts.csswg.org/mediaqueries-4/#media-types
+ * https://drafts.csswg.org/mediaqueries/#media-types
+ * https://www.w3schools.com/CSSref/css3_pr_mediaquery.asp
+ *
+ * `aural`      - DEPRECATED: Used for speech and sound synthesizers
+ * `braille`    - DEPRECATED: Used for braille tactile feedback devices
+ * `embossed`   - DEPRECATED: Used for paged braille printers
+ * `handheld`   - DEPRECATED: Used for small or handheld devices
+ * `projection` - DEPRECATED: Used for projected presentations, like slides
+ * `tty`        - DEPRECATED: Used for media using a fixed-pitch character grid, like teletypes and terminals
+ * `tv`         - DEPRECATED: Used for television-type devices
+ */
+
+export const cssDeprecatedMediaTypes = [
+	'aural',
+	'braille',
+	'embossed',
+	'handheld',
+	'projection',
+	'tty',
+	'tv',
+];