+++ src/constants/cssMediaTypes/index.js
@@ -1,33 +1,21 @@
+import { cssDeprecatedMediaTypes } from '../cssDeprecatedMediaTypes';
+
 /**
  * https://developer.mozilla.org/en-US/docs/Web/CSS/Media_Queries/Using_media_queries#Media_types
  * https://drafts.csswg.org/mediaqueries-4/#media-types
  * https://drafts.csswg.org/mediaqueries/#media-types
- * https://www.w3schools.com/CSSref/css3_pr_mediaquery.asp
  *
  * `all`        - Used for all media type devices
- * `aural`      - DEPRECATED: Used for speech and sound synthesizers
- * `braille`    - DEPRECATED: Used for braille tactile feedback devices
- * `embossed`   - DEPRECATED: Used for paged braille printers
- * `handheld`   - DEPRECATED: Used for small or handheld devices
  * `print`      - Used for printers
- * `projection` - DEPRECATED: Used for projected presentations, like slides
  * `screen`     - Used for computer screens
  * `speech`     - Used for speech and sound synthesizers
- * `tty`        - DEPRECATED: Used for media using a fixed-pitch character grid, like teletypes and terminals
- * `tv`         - DEPRECATED: Used for television-type devices
  */
 
 export const cssMediaTypes = [
 	'all',
-	'aural',
-	'braille',
-	'embossed',
-	'handheld',
 	'print',
-	'projection',
 	'screen',
 	'speech',
-	'tty',
-	'tv',
 	// '3d-glasses',
+	...cssDeprecatedMediaTypes,
 ];