+++ src/handlers/atrules/handleAtRuleDescriptor/index.js
@@ -1,0 +1,6 @@
+import { countUsage } from '../../../calculators/countUsage';
+
+export function handleAtRuleDescriptor(decl, report) {
+	report.atRules.descriptors.total++;
+	countUsage(decl.prop.toLowerCase(), report.atRules.descriptors.usage);
+}