+++ src/handlers/handleDeclaration/index.spec.js
@@ -29,6 +29,7 @@ describe('Module: handleDeclaration', () => {
 			display: flex !important;
 			position: sticky;
 			z-index: -1;
+			margin: 0 auto;
 			float: none;
 			border-radius: 1px 2px 3px 4px;
 			letter-spacing: 0.3em;
@@ -78,7 +79,7 @@ describe('Module: handleDeclaration', () => {
 	describe('Handling declarations', () => {
 		describe('declarations.total', () => {
 			it('should be counted correctly', () => {
-				expect(report.declarations.total).toBe(19);
+				expect(report.declarations.total).toBe(20);
 			});
 		});
 
@@ -96,7 +97,7 @@ describe('Module: handleDeclaration', () => {
 
 		describe('declarations.length.total', () => {
 			it('should be counted correctly', () => {
-				expect(report.declarations.length.total).toBe(452);
+				expect(report.declarations.length.total).toBe(466);
 			});
 		});
 
@@ -122,7 +123,7 @@ describe('Module: handleDeclaration', () => {
 	describe('Handling properties', () => {
 		describe('properties.total', () => {
 			it('should be counted correctly', () => {
-				expect(report.properties.total).toBe(19);
+				expect(report.properties.total).toBe(20);
 			});
 		});
 
@@ -132,9 +133,21 @@ describe('Module: handleDeclaration', () => {
 			});
 		});
 
+		describe('properties.uniqueRatio', () => {
+			it('should be counted correctly', () => {
+				expect(report.properties.uniqueRatio).toBe(0);
+			});
+		});
+
 		describe('properties.shorthands', () => {
 			it('should be counted correctly', () => {
-				expect(report.properties.shorthands).toBe(6);
+				expect(report.properties.shorthands).toBe(7);
+			});
+		});
+
+		describe('properties.shorthandsRatio', () => {
+			it('should be counted correctly', () => {
+				expect(report.properties.shorthandsRatio).toBe(0);
 			});
 		});
 
@@ -189,6 +202,7 @@ describe('Module: handleDeclaration', () => {
 					position: 1,
 					'z-index': 1,
 					float: 1,
+					margin: 1,
 					'border-radius': 1,
 					'letter-spacing': 1,
 					all: 4,
@@ -201,26 +215,33 @@ describe('Module: handleDeclaration', () => {
 			});
 		});
 
-		describe('properties.explicitDefaultingKeywords.total', () => {
+		describe('properties.autoKeyword', () => {
 			it('should be counted correctly', () => {
-				expect(report.properties.explicitDefaultingKeywords.total).toBe(4);
+				expect(report.properties.autoKeyword).toBe(1);
 			});
 		});
 
-		describe('properties.explicitDefaultingKeywords.unique', () => {
+		describe('properties.inheritKeyword', () => {
 			it('should be counted correctly', () => {
-				expect(report.properties.explicitDefaultingKeywords.unique).toBe(0);
+				expect(report.properties.inheritKeyword).toBe(1);
 			});
 		});
 
-		describe('properties.explicitDefaultingKeywords.usage', () => {
+		describe('properties.initialKeyword', () => {
 			it('should be counted correctly', () => {
-				expect(report.properties.explicitDefaultingKeywords.usage).toStrictEqual({
-					inherit: 1,
-					initial: 1,
-					revert: 1,
-					unset: 1,
-				});
+				expect(report.properties.initialKeyword).toBe(1);
+			});
+		});
+
+		describe('properties.revertKeyword', () => {
+			it('should be counted correctly', () => {
+				expect(report.properties.revertKeyword).toBe(1);
+			});
+		});
+
+		describe('properties.unsetKeyword', () => {
+			it('should be counted correctly', () => {
+				expect(report.properties.unsetKeyword).toBe(1);
 			});
 		});
 	});