+++ src/handlers/handleUnits/index.js
@@ -11,9 +11,10 @@ function walkNodes(nodes, decl, report, options) {
 			const unit = node.unit.toLowerCase();
 			const valueWithUnit = node.value + unit;
 
-			if (cssUnits.includes(unit)) {
-				report.units.total++;
+			report.units.total++;
+			countUsage(unit, report.units.usage);
 
+			if (cssUnits.includes(unit)) {
 				/** Count negative margins */
 				if (
 					decl.prop.startsWith('margin') &&
@@ -31,11 +32,13 @@ function walkNodes(nodes, decl, report, options) {
 					report.units.excessive.total++;
 					countUsage(valueWithUnit, report.units.excessive.usage);
 				}
-
-				countUsage(unit, report.units.usage);
+			} else {
+				report.units.unknown.total++;
+				countUsage(unit, report.units.unknown.usage);
 			}
 		}
 
+		/* istanbul ignore else */
 		if (node.type === 'func' && Array.isArray(node.nodes)) {
 			walkNodes(node.nodes, decl, report, options);
 		}
@@ -48,10 +51,12 @@ export function handleUnits(decl, report, options) {
 	try {
 		const ast = parser(decl.value).parse();
 
+		/* istanbul ignore else */
 		if (isSafeAst(ast)) {
 			walkNodes(ast.nodes[0].nodes, decl, report, options);
 		}
 	} catch (err) {
+		/* istanbul ignore next */
 		/* eslint-disable-next-line no-console */
 		console.log(`'postcss-values-parser' module error\n${err}`);
 	}