+++ src/index.spec.js
@@ -15,6 +15,7 @@ describe('Module: scancss', () => {
 			display: inline-block;
 			position: fixed;
 			z-index: '';
+			margin: 0 auto;
 			border: 1px solid #123123;
 			width: 40px;
 			height: 40px;
@@ -94,11 +95,11 @@ describe('Module: scancss', () => {
 			report = null;
 		});
 
-		describe('.styleSheetSize', () => {
+		describe('.size', () => {
 			it('should be counted correctly', () => {
-				expect(report.styleSheetSize).toStrictEqual({
-					source: 1010,
-					gzipSource: 487,
+				expect(report.size).toStrictEqual({
+					source: 1029,
+					gzipSource: 499,
 				});
 			});
 		});
@@ -113,8 +114,8 @@ describe('Module: scancss', () => {
 						shortest: 32,
 						average: 32,
 					},
-					sizeRatio: 0.031683168316831684,
-					sizeRatioPercent: '3.17%',
+					sizeRatio: 0.031098153547133137,
+					sizeRatioPercent: '3.11%',
 				});
 			});
 		});
@@ -130,6 +131,14 @@ describe('Module: scancss', () => {
 						unique: 1,
 						usage: {
 							foobar: 1,
+						},
+					},
+					descriptors: {
+						total: 2,
+						unique: 2,
+						usage: {
+							'font-family': 1,
+							src: 1,
 						},
 					},
 					usage: {
@@ -163,6 +172,11 @@ describe('Module: scancss', () => {
 					types: {
 						total: 1,
 						unique: 1,
+						deprecated: {
+							total: 0,
+							unique: 0,
+							usage: {},
+						},
 						usage: {
 							screen: 1,
 						},
@@ -171,6 +185,11 @@ describe('Module: scancss', () => {
 						total: 1,
 						unique: 1,
 						prefixed: 0,
+						deprecated: {
+							total: 0,
+							unique: 0,
+							usage: {},
+						},
 						usage: {
 							'max-width': 1,
 						},
@@ -304,8 +323,8 @@ describe('Module: scancss', () => {
 							[0, 1, 0],
 						],
 					},
-					sizeRatio: 0.100990099009901,
-					sizeRatioPercent: '10.1%',
+					sizeRatio: 0.09912536443148688,
+					sizeRatioPercent: '9.91%',
 					usage: {
 						'.selector': 3,
 						'::-moz-placeholder': 1,
@@ -320,19 +339,19 @@ describe('Module: scancss', () => {
 		describe('.declarations', () => {
 			it('should be counted correctly', () => {
 				expect(report.declarations).toStrictEqual({
-					total: 20,
-					unique: 18,
-					uniqueRatio: 0.9,
+					total: 21,
+					unique: 19,
+					uniqueRatio: 0.9047619047619048,
 					important: 0,
-					averagePerRule: 2,
+					averagePerRule: 2.1,
 					length: {
-						total: 319,
+						total: 333,
 						longest: 32,
 						longestDeclaration: 'animation:infinite ease 1s fade',
-						average: 15.95,
+						average: 15.86,
 					},
-					sizeRatio: 0.31584158415841584,
-					sizeRatioPercent: '31.58%',
+					sizeRatio: 0.3236151603498542,
+					sizeRatioPercent: '32.36%',
 					inAtRules: {
 						'-webkit-keyframes': 2,
 						keyframes: 2,
@@ -342,6 +361,7 @@ describe('Module: scancss', () => {
 						'display:inline-block',
 						'position:fixed',
 						'z-index:\'\'',
+						'margin:0 auto',
 						'border:1px solid #123123',
 						'width:40px',
 						'height:40px',
@@ -365,20 +385,22 @@ describe('Module: scancss', () => {
 		describe('.properties', () => {
 			it('should be counted correctly', () => {
 				expect(report.properties).toStrictEqual({
-					total: 20,
-					unique: 11,
-					shorthands: 4,
+					total: 21,
+					unique: 12,
+					uniqueRatio: 0.5714285714285714,
+					shorthands: 5,
+					shorthandsRatio: 0.23809523809523808,
 					prefixed: 0,
-					unitless: 13,
+					unitless: 14,
 					resetsViaAll: 0,
 					negativeMargins: 0,
 					anonymousReplacedElements: 0,
 					performanceHacks: {},
-					explicitDefaultingKeywords: {
-						total: 0,
-						unique: 0,
-						usage: {},
-					},
+					autoKeyword: 1,
+					inheritKeyword: 0,
+					initialKeyword: 0,
+					revertKeyword: 0,
+					unsetKeyword: 0,
 					engineTriggers: {
 						composite: {
 							blink: 16,
@@ -402,6 +424,7 @@ describe('Module: scancss', () => {
 					usage: {
 						animation: 2,
 						'background-color': 2,
+						margin: 1,
 						border: 1,
 						color: 3,
 						display: 1,
@@ -683,10 +706,12 @@ describe('Module: scancss', () => {
 		describe('.functions', () => {
 			it('should be counted correctly', () => {
 				expect(report.functions).toStrictEqual({
-					total: 0,
-					unique: 0,
+					total: 2,
+					unique: 1,
 					prefixed: 0,
-					usage: {},
+					usage: {
+						url: 2,
+					},
 				});
 			});
 		});
@@ -751,6 +776,11 @@ describe('Module: scancss', () => {
 						s: 2,
 					},
 					excessive: {
+						total: 0,
+						unique: 0,
+						usage: {},
+					},
+					unknown: {
 						total: 0,
 						unique: 0,
 						usage: {},